Skyscrapers v0.8
The player can select to fill in a tile, or add red or green hints to tile to help them solve the puzzle. The puzzle and number buttons now have their position and scale based off of object within the canvas UI, rather than having to input those values in the inspector. The game UI now scales with screen size and aspect ratio.