Skyscrapers v0.13
Added a shop and a settings menu, both accessible from the main menu. The shop, settings, and tutorial buttons now have a unique design. The background of the main menu has a unique design, which is an expanded version of the cityscapes logo. The game now tracks how many coins a player has collected, which show up in the shop. Several settings with boolean toggles can now be interacted with using the settings or shop menus.