Cityscapes v0.15
All numbers used in the game are interations of the same 10 digits now. Different colorations are now applied through a filter on those digits. Redesigned the 1 digit to be more recognizable as a 1. The coin amount on the Shop menu no longer uses a generic text box. Added a fade-in and fade-out feature to the main menu and the puzzle scenes.