Cityscapes v0.16
Added all letters as separate objects. Main menu background png no longer includes letters, and the letters are instead added as objects on top of the background. Added a background for the shop, which also serves as the background for the settings. Added a basic background for the puzzle scene. Fixed a bug in which applied notes were giant.