Cityscapes v0.18
Buttons on the puzzle screen now have 2 components - a background color and an overlaid icon. The colors for selected and disselected buttons, as well as the icons themselves are now properties of the skin class. Improved the "you win" popup. The black fade-in/fade-out overlay now covers the entirety of the game screens.