Cityscapes v0.19
Updated the Shop and Settings interfaces to include a scrollable menu. Additionally, purchases can be made in the shop (with no coin cost at the moment), which unlock features for the player. These features can be toggled in the Settings screen. Features that have not yet been purchased do not show up in the Settings screen. Lastly, the Settings and Shop backgrounds have been updated to provide a backdrop for the scrollable interfaces.