Cityscapes v0.27
Brown buttons in the upgrade shop no longer turn grey when they have been purchased, and the grey dash has been replaced with a blue check mark. This is to make the layout and coloration of the shop menu more consistent and recognizable. All buttons in the shop and on the main menu have had their images replaced with multiple white button segments, which can be colored separately and dynamically. This is to make it easier to change the colors of these buttons, and also to allow for skin customization to affect the button colors in the shop and main menu screens. Moved some old art assets to a new "Old Art" folder, which will hold art who are currently not in use, but whose destiny has not yet been determined. Aside from the "Menu" button in the shop, both the main menu and the shop scenes are complete at this time!