Cityscapes v0.31
Updated the win popup to have a custom border, similar to the border around the shop scene. Updated the Settings scene to be more like the shop scene. Specifically, the toggle buttons for the settings scene use the new button format, with some overlaid text showing if the unlock is toggled on or off. Split the tutotiral into a separate canvas from the main puzzle canvas, but both are still a part of the same scene. Updated the tutoiral look to include the word "TUTORIAL" on the top, as well as slightly shrinking the puzzle and building number components. The tutorial text box now has the same border layout as the shop and win popup.

At this moment, I have decided to only have the one tutorial for the game, as the unlockables are simple enough to be explained with the relevant blurbs in the shop scene. The tutorial's text and order have not yet been updated, but will be soon.