Cityscapes v0.40
Added the "show/hide purchased upgrades" toggle in the settings. Also changed the tutorial so that it no longer uses the current skin's colors for the menu button. Also cleaned up the shop, settings, and credits scene backgrounds for the space skin. All of the backgrounds for the space skin should use pixel-perfect, non-blurry scaling for the planets and backdrop.