Cityscapes v0.41
Fixed a bug in the shop that would expand the scroll view to be too large when purchasing an upgrade. Added the ability to remember progress on a puzzle when exiting to the main menu, and you can return to the puzzle with all of your progress retained. This does not yet work when closing out of the app.