Cityscapes v0.44
Added the Turquoise skin, mainly for testing purposes. The shop can now support different tiers of prices. The shop can now find a skin's cost and execute its purchase purely based off of the button's name in the shop, so adding new skins to the shop is very easy. Adding skins to the settings has also been made easier by changing the name of the skin selection text to match the button's name in the settings. The credits scene text has been updated to reflect the mountain and space skin assets. The "return to city" / "abandon city" interface has been updated with helpful text, as well as a background whose color updates to match the skin.