Cityscapes v0.46
Added the ability for different skins to have different building designs and color schemes for the backdrop behind the city-in-progress popup. Each skin has one building design, whose relative size varies between 50% and 100% depending on the building size number. The basic skin's building has been created, which may or may not be a temporary design.