Cityscapes v0.48
When a puzzle is left in progress, the main menu now only features the puzzle-in-progress popup with a new grey semitransparent backdrop behind it. Also the tutorial no longer uses button colors from the current skin, and instead uses the default skin colors. Finally, the game is saved whenever the player makes any action in the puzzle scene, and also when the player makes any purchase from the shop or changes any upgrade from the settings.