Cityscapes v0.49
Improved the puzzle background for the Space Skin. Updated the app icon. Buttons can now be clicked during fade-in to a scene, and the click function will execute after the fading in process is complete. The in-progress popup on the main menu now features the appropriate city art for the size of the city that you were working on, and the art will change to match the skin. The "Abandon city" button has the same art as the "Return to City" button, but semitransparent. A new icon has been added which is a modified version of the small city button with the yellow CITYSCAPES text from the main menu. This icon will be the icon used on the Google Play Store. The splash screen shown when the app is booting up now shows the same app store icon, now on the same screen as the Unity logo.