Cityscapes v0.50
Removed all unused assets. Centered the adaptive app icon logo. The color of the city streets can now vary with skin. The red borders used in the tutorial are now made to be more blocky and pixelated, and less blurry and rounded. Removed the Purple and Turquoise variants of the Basic Skin from the shop and settings. They still remain in the game assets, as well as all of their relevant art.