Cityscapes v0.52
Improved the road color and building colors for the mountains and space skins. The game now loads into the main menu with a fade-in. The game now saves after you choose to abandon a puzzle. The red border around side hint tiles, used for the tutorial, was previously removed and is now back, now using the updated pixelated-corner red border sprite.