Cityscapes v0.53
Added city art for the Mountains Skin. Updated the Credits Scene to have the same scrolling window as the shop and settings scenes. Added the ability to use the back button on android to change scenes. In the InPuzzle Scene, the main menu button and the back button functionality have been disabled after the user has won their puzzle, and the only way to return to the main menu is via the button on the win popup. Added the Highlight Buildings unlock, which highlights all instances of a specific building height in the puzzle to match the building type the player has already selected. Added more text to the shop scroll view - one chunk of text that appears after the player has purchased every upgrade, and 3 sets of text that appear when the player has purchased all of a single type of upgrade and if they have the "Hide Purchased Upgrades" toggle on.