Cityscapes v0.54
Removed old Basic Skin Art. Removed the "Credits Background" property of skins, as it is no longer used. If the player starts a puzzle and does not take any actions then leaves the puzzle, the puzzle state is not remembered. The main menu has been updated so that if the player returns to the main menu from a puzzle, the "In Progress" popup shows, but the shop, tutorial, and settings buttons also show. Added the centerContent function to GameManager, which will move buttons and the puzzle around to be centered in the screen if only some buttons are active. centerContent is currently disabled, but will be functional when enabled. Improved the tutorial text so that adding text while retaining the old text looks less choppy. Updated the company name to be Cole Hilscher, and changed the version number to 1.0. Removed old unused code that has been commented out.