Cityscapes v0.57
Added the ability to go to the shop from the win screen in a puzzle. Also added a way to display the number of coins the player currently has on the win screen. Added a giant tutorial button on the main menu, which appears when the player opens the game for the first time. This button only goes away when the player completes the tutorial. While this button is showing, no other buttons on the main menu are visibile. Added a temporary button in the shop to make the tutorial button show again. Sorted the toggle buttons in the Setting scene, which are now grouped by cosmetics, visual upgrades, buttons, city upgrades, and other settings. When any button from one of these categories is present, a relevant title is also shown above the group. Changed the Android Min Sdk Version to 28, as required by Google Play for publishing.