Cityscapes v0.63
Added the Erase button, which does not need to be unlocked in order to use it in a puzzle. Added the Eraser icon, and also moved the build, note1, and note2 icons out from their respective Skin folders. These icons are now static between all skins. The "return to city" button on the main menu now functions even if the player has the associated city size locked from settings. Changed the company name to High Gravity Games in preparation for release. This version of the game is now Version 1.0 and to be used in the first Beta Release.