Text and scrollview improvements
Cleaned up the descriptive text in the shop and the settings scenes.

Improved the scrollview in the shop so that if the expanded item is on the bottom edge of the viewport, the scroll view auto-scrolls down to include the full expanded item size.

Separated the credits text into different gameobjects to make it easier to add more credits in the future.