Prep for massive cities
Reorganized the layout of the puzzle scene, to make room for larger / more number buttons. Added an icon to the clear puzzle button.

Did all of the prep required for 7x7 puzzles. All that remains to reorganize the home screen to accommodate for the new puzzle button, add massive city icons for the current skins, and improve the puzzle generation algorithm.