New icons, win popup improvements, and improved city art
Tutorial and shop icons are now using nicer artwork. This applies to the main menu, settings, and puzzle scenes.
Reorganized the win popup to use the new home / shop / settings buttons, and made more room for the new puzzle button to be larger.
Created prefabs for the city size artwork, and moved the objects around within the prefabs to make positioning within scenes easier.