Fixed an error if the last skin was Basic
If the last skin active was the Basic skin before updating the game version, the skin is switched to Rural instead.
Also removed the "Skin" suffix from the skin names in the skin selection in settings. Renamed Basic to Rural in the Credits references, and reorganized the skin list in the credits.