More tutorial improvements
Improved the tutorial text, and changed the order that buildings get filled in for the second city. Now the player is guided to do a street that has all 3 buildings in order from shortest to tallest. This is easier to explain than the previous iteration, which had the player placing a 2-story building on one end of a street that had a 3-story building on the other end.