Removed a lot of unused files
If you want to go back later and add more skins, a lot of assets that could be reworked into skins are being removed with this change. They are in the Assets/2d art/test skin stuff folder, and almost all of them come free from itch.io (for finding original artists to credit in the future).