Shop, Settings, and Credits Scene improvements
Added a scroll bar to the shop, settings, and credits scenes, and adjusted some font sizing to not overlap with the scroll bar. Added spacing on the urls listed in each credit, to improve readability. Adjusted the text that displays in the shop once you have unlocked all upgrades and all skins.