Skin cost changes
The first skin a player buys is free, with the cost of the skins getting incrementally more expensive as the player unlocks more skins. 0 > 10 > 30 > 60 > 100, and that pattern repeats forever and ever. Added some text to the shop to explain this behavior to the player.
Fixed an issue where buying a new skin would not fully apply it to everything in the shop scene right away.