Added game version tracking
StaticVariables now records the game version number. On launching the game, the version number is checked, and if it is out of date, an update function is run to update the version number.

As part of this, when updating to v2.1, the player is refunded coins if they bought skins before updating, to account for the new skin pricing.