Aspect Ratio expansion pt 1
Made the rural skin wider and taller, from the old fixed size of 1440x2960 to a new bigger size of 2000x3200. This will allow the game window to adjust to various aspect ratios.
The end behavior is the same as before, but instead of having black bars on the side or top of the screen, that empty space has been filled with more skin background art.
To accommodate this change, the size of many UI elements has been adjusted.
This is only part 1 of the change, because all of the other skins have to be modified as well.