Cleaned up the settings scene
Instead of individual buttons for toggling each purchased unlock, now there is one button that opens a dropdown where all unlocks are contained, similar to the skin-switching dropdown.
Because of this, the settings scene is now a lot less cluttered.
Alongside this, the descriptions of some unlocks have been reworded and updated in the shop scene.