Simplified button toggling in the puzzle scene
Previously, there were multiple parent gameobjects that contained the build, erase, notes1, and notes2 buttons, and multiple parents that contained the undo, redo, removeall, and clearpuzzle buttons. Which parent gameobject was set to active depended on which buttons were supposed to be usable in the puzzle.
Now, there is one parent gameobject for each button grouping, and they use a horizontal layout group to dynamically reposition the child buttons depending on which are supposed to be usable.
In effect, the gamemanager script code has been decluttered significantly, and there are far fewer gameobjects present in the inpuzzle scene. Game functionaly has not changed.