Added the fill notes unlock
Like other unlocks, it can be purchased from the shop and toggled in settings.
With this feature enabled, any time the player has a note type selected and there are no notes of that type in the city, the remove all button can be pushed to fill all empty spaces with the selected note.