Added Skin Applicators
Every scene now has a list of all objects which need to have their colors adjusted to match the current skin. These Skin Applicator objects are contained in a list belonging to each scene manager, and every time a skin needs to be loaded, those applicator objects are iterated through and the updated skin is applied to each.
A lot of ugly code has been cleaned up as a result, and many transform.Find and object.GetComponent calls have been removed, which should improve runtime.